Would you like some fries with that cheeseburger, sir?
HHHH THE PUSH TO THE STEAM WORKSHOP IS COMING SOON. AS ALWAYS, PLEASE REMEMBER THIS:

"A delayed update is always good. A rushed one is forever bad."

NOTICE THAT LORE! IS IN ACTIVE DEVELOPMENT IN SOME PARTS.

That's why sh*t's moving a bit slowly! Don't worry, everything is fine, and is being tested every 5 seconds.